Remove CircleCI configuration
It's been broken for a while, and when I tried to fix it, TensorFlow installation failed while grpc was being compiled, with an internal gcc error that sounds like an out-of-memory error. Test runners with more RAM are only available with a paid plan, so ugh, for now just remove it all.